For consistency, let's use decimals not doubles

+++ CSharp.Nixill.GTFS/src/GTFS/Entities/IDEntities/Level.cs
@@ -5,7 +5,7 @@ namespace Nixill.GTFS.Entities
 {
   public class Level : GTFSIdentifiedEntity
   {
-    public double LevelIndex => Properties.GetDouble("level_index");
+    public decimal LevelIndex => Properties.GetDecimal("level_index");
     public string LevelName => Properties["level_name"];
 
     public Level(GTFSPropertyCollection properties) : base(properties, "level_id") { }
